### 2017-07-09 #308 @whedon
[PRE REVIEW]: USAboundaries: Historical and Contemporary Boundaries of the United States of America
**Submitting author:** @lmullen (<a href="http://orcid.org/0000-0001-5103-6917">Lincoln Mullen</a>)
**Repository:** <a href="https://github.com/ropensci/USAboundaries" target ="_blank">https://github.com/ropensci/USAboundaries</a>
**Version:** v0.3.0
