### 2017-07-10 #314 @arfon
@openjournals/ropensci - could you link to the rOpenSci onboarding thread?
