### 2017-07-10 #314 @lmullen
@arfon I've tagged everything 0.3.0 and create a GitHub release. Here is the Zenodo DOI.

DOI: 10.5281/zenodo.825218
