### 2017-07-10 #314 @arfon
> This package was developed under rOpenSci but has not been peer-reviewed through rOpenSci onboarding. (We plan to make it easier to distinguish between the two with some new badging and metadata soon).

Ah OK. @karthik - looks like we need a reviewer for thus submission then...
