### 2017-07-27 #314 @AndySouth
Thanks @karthik, happy to do this. I'm travelling for a week will get to it after then. On quick inspection I couldn't find reviewers instructions, e.g. turnaround time, can you point me to ?
