### 2017-07-27 #314 @karthik
Thanks @AndySouth. Guidelines are here: http://joss.theoj.org/about#reviewer_guidelines
If you are able to accept and complete the review within two weeks after you're back, I would really appreciate it. Please let me know if that's ok and I'll start off the review process (which will close this issue and create a new one with a checklist for you to work through).
