### 2017-08-20 ropensci/USAboundaries#35 @AndySouth
add states argument to `us_cities()` and `us_zipcodes()`
`us_cities()` and `us_zipcodes()` don't yet offer states argument as indicated in the readme. Can they be added ?
