### 2017-08-20 ropensci/USAboundaries#36 @AndySouth
Can us_cities() return an sf object ?
Is there a reason that us_cities returns a dataframe and not an sf points object ? Returning as an sf object would help us_cities() being used in combination with us_states(). I tried sf::plot(st_geometry(us_cities())) expecting it to work the same as us_states().

Also can use of us_cities() be added to the readme ?
