### 2017-08-20 ropensci/USAboundaries#37 @AndySouth
Can potential for joining county fields from us_cities() and us_counties() be improved ?
`us_cities() `returns a dataframe with a fields called 'county' and 'county_name'. These don't seem quite to link to the 'name' field from `us_counties()`. It would be good to have an indication of how best these datasets could be used together, or to have an issue for how this can be improved in future. e.g. for Alaska, in the cities dataset some county names have Borough or census area appended, and some e.g. Anchorage have a very different name 'THIRD JUDICIAL DIVISION'.

```
counties_ak <- us_counties(states = "Alaska")
unique(counties_ak$name)
```

 [1] "Bristol Bay"           "Skagway"               "Ketchikan Gateway"     "Nome"
 [5] "Yakutat"               "Kenai Peninsula"       "Lake and Peninsula"    "Juneau"
 [9] "Bethel"                "Kodiak Island"         "Aleutians East"        "Anchorage"
[13] "North Slope"           "Aleutians West"        "Southeast Fairbanks"   "Fairbanks North Star"
[17] "Sitka"                 "Denali"                "Prince of Wales-Hyder" "Yukon-Koyukuk"
[21] "Haines"                "Petersburg"            "Valdez-Cordova"        "Matanuska-Susitna"
[25] "Dillingham"            "Northwest Arctic"      "Hoonah-Angoon"         "Kusilvak"
[29] "Wrangell"

`dplyr::filter(uscities,state=='AK') %>% dplyr::distinct(county)`

 1         THIRD JUDICIAL DIVISION
 2             North Slope Borough
 3              Bethel Census Area
 4        FOURTH JUDICIAL DIVISION
 5         Kenai Peninsula Borough
 6         FIRST JUDICIAL DIVISION
 7           Kodiak Island Borough
 8        Northwest Arctic Borough
 9               NORTHERN DISTRICT
10       Matanuska-Susitna Borough
11 Wrangell-Petersburg Census Area
12                   Sitka Borough
13      Aleutians West Census Area
14      Valdez-Cordova Census Area
