### 2017-08-20 ropensci/USAboundaries#38 @AndySouth
`state_plane()` add example of plotting to the examples section. 
Can be taken from the readme.
