### 2017-08-20 ropensci/USAboundaries#39 @AndySouth
add references to the JOSS paper submission
Add to paper.md a list of key references including a link to the software archive. The former can be copied from the Readme.
