### 2017-08-20 ropensci/USAboundaries#40 @AndySouth
Add an example of a real world analysis.
Add to the readme an example of how the package data can be joined to some other data and mapped. This can help new users see how they can use with their own data.
From JOSS guideline to reviewers : Example usage : The authors should include examples of how to use the software (ideally to solve real-world analysis problems).
