### 2017-08-20 ropensci/USAboundaries#41 @AndySouth
statement of need can be strengthened in the readme
I am sure this is a hugely useful package and that it could help users with lots of spatial analyses. Could a brief discussion of potential uses be added to the readme ? Currently the readme might suggest to a naive reader that it can only be used for plotting.
