### 2017-08-20 #314 @AndySouth
@lmullen @karthik

I've worked through the checklist and submitted some issues to the project repo.

I think this is a super-useful package, should I write a summary paragraph here too ?

[#](https://github.com/ropensci/USAboundaries/issues/41)
[#](https://github.com/ropensci/USAboundaries/issues/40)
[#](https://github.com/ropensci/USAboundaries/issues/39)
[#](https://github.com/ropensci/USAboundaries/issues/38)
[#](https://github.com/ropensci/USAboundaries/issues/37)
[#](https://github.com/ropensci/USAboundaries/issues/36)
[#](https://github.com/ropensci/USAboundaries/issues/35)
