### 2017-08-22 #314 @karthik
Thanks @AndySouth! We really appreciate the detailed review. No summary necessary here since you've linked to the issues (tip for next time, mentioning `joss-reviews/314` in the issues you open will automatically link them here with the title). I will wait for @lmullen to respond to these comments (push an update or respond with a reason) and once these are addressed, I'll ping you for another really quick check before proceeding with the acceptance.

Thanks again for sharing your time and expertise, Andy!
