### 2017-08-24 #314 @lmullen
Thanks for the review, @AndySouth. I will go over these issues as soon as I can and get back to you both with a summary response and changes to the package/paper.
