### 2017-10-13 9484492
Filter by state_terr rather than state_name
The counties that became Tennessee but were previously in North Carolina are listed under North Carolina in `state_terr` (which is correct) but under Tennessee in `state_name`, which I guess I understand why they did that but it's not correct.
