### 2017-10-16 ba3c86f
Fix bug when state_terr field does not exist
