+++ R/filters.R
@@ -7,8 +7,14 @@ filter_by_states <- function(shp, states) {
   if (is.null(states)) return(shp)
   stopifnot(is.character(states))
   states <- tolower(states)
-  filter <- (tolower(shp$state_terr) %in% states) |
-    (tolower(shp$state_abbr) %in% states)
+  # Use state_terr for historical, otherwise statename
+  if (!is.null(shp$state_terr)) {
+    filter <- (tolower(shp$state_terr) %in% states) |
+      (tolower(shp$state_abbr) %in% states)
+  } else {
+    filter <- (tolower(shp$state_name) %in% states) |
+      (tolower(shp$state_abbr) %in% states)
+  }
   if (sum(filter, na.rm = TRUE) < 1) stop("No matches found for those states.")
   shp[filter, ]
 }
