### 2018-03-20 efc4c9c
Add stronger statement of need to README
Fixes #41
