+++ README.md
@@ -1,11 +1,34 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-USAboundaries: Historical and Contemporary Boundaries of the United States of America
--------------------------------------------------------------------------------------
 
-[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/USAboundaries)](http://cran.r-project.org/package=USAboundaries) [![Build Status](https://travis-ci.org/ropensci/USAboundaries.png?branch=master)](https://travis-ci.org/ropensci/USAboundaries) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/USAboundaries?branch=master)](https://ci.appveyor.com/project/ropensci/USAboundaries) [![Coverage Status](https://img.shields.io/codecov/c/github/ropensci/USAboundaries/master.svg)](https://codecov.io/github/ropensci/USAboundaries?branch=master)
+## USAboundaries: Historical and Contemporary Boundaries of the United States of America
 
-This R package includes contemporary state, county, and Congressional district boundaries, as well as zip code tabulation area centroids. It also includes historical boundaries from 1629 to 2000 for states and counties from the Newberry Library's [Atlas of Historical County Boundaries](http://publications.newberry.org/ahcbp/), as well as historical city population data from Erik Steiner's "[United States Historical City Populations, 1790-2010](https://github.com/cestastanford/historical-us-city-populations)." The package has some helper data, including a table of state names, abbreviations, and FIPS codes, and functions and data to get [State Plane Coordinate System](https://en.wikipedia.org/wiki/State_Plane_Coordinate_System) projections as EPSG codes or PROJ.4 strings.
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/USAboundaries)](http://cran.r-project.org/package=USAboundaries)
+[![Build
+Status](https://travis-ci.org/ropensci/USAboundaries.png?branch=master)](https://travis-ci.org/ropensci/USAboundaries)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/ropensci/USAboundaries?branch=master)](https://ci.appveyor.com/project/ropensci/USAboundaries)
+[![Coverage
+Status](https://img.shields.io/codecov/c/github/ropensci/USAboundaries/master.svg)](https://codecov.io/github/ropensci/USAboundaries?branch=master)
+
+This R package includes contemporary state, county, and Congressional
+district boundaries, as well as zip code tabulation area centroids. It
+also includes historical boundaries from 1629 to 2000 for states and
+counties from the Newberry Library’s [Atlas of Historical County
+Boundaries](http://publications.newberry.org/ahcbp/), as well as
+historical city population data from Erik Steiner’s “[United States
+Historical City
+Populations, 1790-2010](https://github.com/cestastanford/historical-us-city-populations).”
+The package has some helper data, including a table of state names,
+abbreviations, and FIPS codes, and functions and data to get [State
+Plane Coordinate
+System](https://en.wikipedia.org/wiki/State_Plane_Coordinate_System)
+projections as EPSG codes or PROJ.4 strings.
+
+This package can serve a number of purposes. The spatial data can be
+joined to any other kind of data in order to make thematic maps. Unlike
+other R packages, this package also contains historical data for use in
+analyses of the recent or more distant past.
 
 ### Installation
 
@@ -13,36 +36,52 @@ You can install this package from CRAN.
 
     install.packages("USAboundaries")
 
[…]
