+++ Makefile
@@ -1,0 +1,9 @@
+.PHONY : docs deploy-docs
+
+docs :
+	Rscript -e "pkgdown::clean_site(); pkgdown::build_site(run_dont_run = TRUE)"
+
+deploy-docs :
+	cp -r tools docs/
+	rsync --checksum --delete -avz docs/* reclaim:~/www/lincolnmullen.com/software/usaboundaries/
+
