+++ _pkgdown.yml
@@ -1,11 +1,10 @@
-url: http://ropensci.github.io/USAboundaries/
+url: http://lincolnmullen.com/software/tokenizers
+
+template:
+  ganalytics: "UA-25121492-1"
+  params:
+    bootswatch: united
+
 authors:
   Lincoln Mullen:
-    href: http://lincolnmullen.com
-  Jordan Bratt:
-    href: http://jfbratt.org/
-template:
-  params:
-    bootswatch: lumen
-    ganalytics: UA-25121492-1
-
+    href: "http://lincolnmullen.com"
