### 2018-03-20 4f2ee3f
Consistent format for README
