+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: USAboundaries
 Type: Package
 Title: Historical and Contemporary Boundaries of the United States of America
-Version: 0.3.0
+Version: 0.3.0.9000
 Date: 2017-07-05
 Authors@R: c(person("Lincoln", "Mullen",
     email = "lincoln@lincolnmullen.com",
