+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: http://lincolnmullen.com/software/tokenizers
+url: http://lincolnmullen.com/software/usaboundaries
 
 template:
   ganalytics: "UA-25121492-1"
