### 2018-03-20 4760015
Make `us_cities()` return an sf object
Fixes #36

Most of the work was done in USAboundariesData: https://github.com/ropensci/USAboundariesData/commit/c497fbb26ce78b9a768594abe2e5e072be0c5382
