### 2018-03-20 ropensci/USAboundaries#41 @lmullen
I have expanded on this in the README and will do so shortly in the revised JOSS paper.
