### 2018-03-20 ropensci/USAboundaries#36 @lmullen
This was changed in the USAboundariesData package and `USAboundaries::us_cities()` now returns an sf data.frame.
