### 2018-03-20 5bf1c6d
Example of plotting using state plane projections
Fixes #38
