+++ R/state_plane.R
@@ -22,6 +22,14 @@
 #' state_plane(state = "MA", plane_id = "island", type = "epsg")
 #' state_plane(state = "MA", plane_id = "island", type = "proj4")
 #'
+#' # Show the difference made by a state plane projection
+#' if (require(sf)) {
+#'   va <- us_states(states = "VA", resolution = "high")
+#'   plot(st_geometry(va), graticule = TRUE)
+#'   va <- st_transform(va, state_plane("VA"))
+#'   plot(st_geometry(va), graticule = TRUE)
+#' }
+#'
 #' @export
 state_plane <- function(state, plane_id = NULL, type = c("epsg", "proj4")) {
   type <- match.arg(type)
