### 2018-03-20 ropensci/USAboundaries#38 @lmullen
Took the example from the README as suggested.
