### 2018-03-20 ropensci/USAboundaries#37 @lmullen
@AndySouth I appreciate the suggestion but I don't think that this is a step that I can take for this package. The reason is that all the data provided by this package comes from other sources (as cited). I have made these data sets as interoperable as possible by casting them all to sf objects, standardizing column names, providing them with the same CRS, and providing a consistent interface. But since I am reporting other people's data I have been reluctant to change the data itself. In this particular case, too, verifying the county data in the CESTA cities data set would be as time consuming as gathering the data in the first place. And doing so would run counter to the historical impetus of the package. Cities rarely move, but county boundaries change all the time.

Ideally this would be possible using the FIPS codes for counties and for places. But while the county FIPS identifier is a combination of a state FIPS code and a county code, the place FIPS identifier is a combination of the state FIPS code and a place code. No county information is embedded in the FIPS place ID. So the difficulty isn't really with this package but with the data model of FIPS identifiers.
