### 2018-03-20 ec38bad
Let us_cities() be filtered by states
Fixes #35
