+++ tests/testthat/test-us_cities.R
@@ -14,3 +14,10 @@ test_that("Cities function returns an sf data.frame", {
   expect_is(us_cities(1876), "data.frame")
   expect_is(us_cities(1976), "sf")
 })
+
+test_that("Cities function can be filtered by state", {
+  nc <- us_cities(2010, states = "NC")
+  sc <- us_cities(1876, states = "South Carolina")
+  expect_identical(unique(nc$state_name), "North Carolina")
+  expect_identical(unique(sc$state_abbr), "SC")
+})
