### 2018-03-20 ropensci/USAboundaries#35 @lmullen
I have made `us_cities()` accept a `states` argument. Zip codes can cross state boundaries so that is not possible for `us_zipcodes()`. But I have updated to the README to make it clear that most but not all functions accept the `states` argument.
