+++ DESCRIPTION
@@ -26,7 +26,7 @@ Depends:
 Suggests:
     sf (>= 0.5.1),
     testthat (>= 0.9.1),
-    USAboundariesData (>= 0.3.0),
+    USAboundariesData (>= 0.3.1),
     covr
 RoxygenNote: 6.0.1
 Additional_repositories: http://packages.ropensci.org
