+++ _pkgdown.yml
@@ -8,3 +8,5 @@ template:
 authors:
   Lincoln Mullen:
     href: "http://lincolnmullen.com"
+  Jordan Bratt:
+    href: "http://jfbratt.org"
