### 2018-03-21 9204504
Vignette demonstrating basic analysis
Fixes #40
