+++ vignettes/usaboundaries-sample-analysis.Rmd
@@ -1,0 +1,102 @@
+---
+title: "A sample analysis using USAboundaries"
+author: "Lincoln Mullen"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{A sample analysis using USAboundaries}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+The USAboundaries package provides both historical and contemporary boundaries of the United States. This vignette gives an example of using the package to create maps of Maryland Congressional elections in 2016 (to the 115th Congress) and in 1792 (to the 3rd Congress). We will use the USAboundaries package data for the spatial data. See the documentation for the functions used for citations to the underlying datasets. For the 2016 Congressional votes, we will use [data available on the website](http://history.house.gov/Institution/Election-Statistics/Election-Statistics/) of the U.S. House of Representatives. For the 1792 elections we will use data from the [*Mapping Early American Elections*](http://earlyamericanelections.org/data/) project.
+
+We begin by loading the packages that we need.
+
+```{r message=FALSE}
+library(USAboundaries)
+library(dplyr)
+library(sf)
+library(leaflet)
+```
+
+We can load the voting data for the 115th Congress into `md_votes`. We use this package to get the contemporary (2016) boundaries for Congressional districts in Maryland. They are returned as an `sf` object from the [sf](https://github.com/r-spatial/sf/) package.
+
+```{r}
+md_votes <- read.csv(system.file("extdata", "md-115.csv", package = "USAboundaries"),
+                   colClasses = c("character", "integer", "character", "integer",
+                                  "character", "integer", "integer", "integer",
+                                  "integer", "integer", "integer",  "numeric",
+                                  "numeric", "numeric", "numeric", "numeric"))
+md_districts <- us_congressional(states = "MD")
+```
+
+In order to map the data, we have to join the voting data to the spatial data.
+
+```{r}
+md_115 <- md_districts %>% 
+ left_join(md_votes, by = c("cd115fp" = "district")) 
+```
+
+The data has the percentage of votes for Democrats and Republicans. We are going to map the margin that Republicans had over an even vote, so that we can see both the Democrat and Republican results. For a more careful analysis we would have to take account of third parties and write-in votes, but this will serve our purpose.
+
+```{r}
+md_115$margin_percent <- md_115$percentage_republican - 0.5
+```
+
+We create a color palette to turn the margin of victory into red and blue colors.
+
+```{r}
+palette <- colorBin("RdBu", domain = c(-0.3, 0.3), bins = 7, reverse = TRUE)
+```
+
+We can use the [leaflet](https://rstudio.github.io/leaflet/) package to make the map.
+
+```{r}
+leaflet(md_115) %>% 
+  addTiles() %>% 
+  addPolygons(color = "black",
+              opacity = 1,
+              weight = 1,
+              fillColor = ~palette(margin_percent),
+              fillOpacity = 1,
+              label = ~paste0("District: ", cd115fp))
+```
+
+We can do the same analysis for the 3rd Congress, thanks to the historical data included in the package. In this case, we will have to map counties rather than Congressional districts. The parties of interest are the Federalists and the Democratic-Republicans.
+
+We start by loading and joining the data. 
+
+```{r}
+md_03_votes <- read.csv(system.file("extdata", "md-003.csv", package = "USAboundaries"),
+                        stringsAsFactors = FALSE)
[…]
