+++ NEWS.md
@@ -1,7 +1,7 @@
 # USAboundaries 0.3.1
 
 - New vignette demonstrating the package's functionality (#40).
-- Additions and clarifications to documentation following @AndySouth's suggestions for JOSS peer review (#38, .
+- Additions and clarifications to documentation following @AndySouth's suggestions for JOSS peer review (#38).
 - `us_cities()` now returns an `sf` object rather than a data frame (#36).
 - `us_cities()` gains a `states` argument to match other functions in the package (#35).
 
