### 2018-03-21 51cedb0
Add dependencies from vignette
