+++ DESCRIPTION
@@ -31,7 +31,9 @@ Suggests:
     USAboundariesData (>= 0.3.1),
     covr,
     knitr,
-    rmarkdown
+    rmarkdown,
+    dplyr,
+    leaflet
 RoxygenNote: 6.0.1
 Additional_repositories: http://packages.ropensci.org
 LazyData: true
