### 2018-03-21 f2b62e5
Revise JOSS paper
Fixes #39

Addresses https://github.com/openjournals/joss-reviews/issues/314
