+++ paper.bib
@@ -2,8 +2,8 @@
     title = {USAboundaries: Historical and Contemporary Boundaries of the United States of America},
     author = {Lincoln Mullen and Jordan Bratt},
     year = {2017},
-    note = {R package version 0.3.0},
-    url = {https://github.com/ropensci/USAboundaries},
+    note = {R package version 0.3.1},
+    url = {http://lincolnmullen.com/software/usaboundaries/},
   }
 
 @misc{steiner-cities,
@@ -37,3 +37,10 @@
   year = {1989}
 }
 
+@misc{meae,
+  author = {Mapping Early American Elections project team},
+  title = {Mapping Early American Elections},
+  institution = {Roy Rosenzweig Center for History and New Media, George Mason University},
+  year = {2016--2019},
+  howpublished = {\url{http://earlyamericanelections.org/}}
+}
