### 2018-03-21 #314 @lmullen
@AndySouth Thanks for your thorough review, which I very much appreciate. Please accept my sincere apologies for the undue delay in implementing the requested changes.

I have gone through the issues that were opened on the USAboundaries issue and addressed them all. In all but one case I made the change that was requested. Those issues should have links to the commits and any other discussion so I won't belabor the point here.

I did also revise `paper.md` to elaborate on some points as requested to show a concrete example of a research project.

@karthik Please do let me know if you would like any further changes.
