+++ _pkgdown.yml
@@ -1,9 +1,9 @@
 url: http://lincolnmullen.com/software/usaboundaries
 
 template:
-  ganalytics: "UA-25121492-1"
   params:
     bootswatch: united
+    ganalytics: "UA-25121492-1"
 
 authors:
   Lincoln Mullen:
