### 2018-03-26 ropensci/USAboundaries#37 @AndySouth
@lmullen Thanks for the clear explanation, I now understand that what I naively asked for is not possible within the scope of this project.
