### 2018-03-26 #314 @lmullen
@AndySouth Thanks very much for the review. It has made the package stronger in a number of ways, not least the consistency of using all sf objects. Much appreciated!
