### 2018-03-26 #314 @karthik
Thanks very much for your time and expertise, @AndySouth 🙏
@lmullen I'm happy with the changes and the explanation for the one exception. Please archive your software on Zenodo and post a DOI here. I will proceed with accepting your submission.

Thank you!
