### 2018-03-26 #314 @lmullen
@karthik Here is the link to the Zenodo archive for v0.3.0 (the version reviewed): https://zenodo.org/record/825218

It's linked in `paper.md` but I'm happy to change the way it is listed if there is a standard way to do so for JOSS.
