### 2018-03-30 #314 @arfon
> @arfon Thanks for the chance to fix minor errors in the PDF. I think that I've got everything fixed now. Could you regenerate the PDF please?

Sure thing. I've updated the paper based on what's in `master` on https://github.com/ropensci/USAboundaries
